engine/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.0
pydantic-settings==2.8.1
httpx==0.28.1
langgraph==0.2.74
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
openai==1.72.0
playwright==1.52.0
